requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==2.0.12
chess==1.9.4
chess.com==2.0.3
click==8.1.3
colorama==0.4.6
docopt==0.6.2
exceptiongroup==1.1.1
Flask==2.2.2
frozenlist==1.3.3
google-api-core==2.11.1
google-auth==2.20.0
google-cloud-bigquery==3.11.1
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.25.0
packaging==23.1
pandas==1.5.2
pathspec==0.11.1
pipreqs==0.4.13
platformdirs==3.6.0
pluggy==1.0.0
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pytest==7.3.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.28.0
rsa==4.9
six==1.16.0
tomli==2.0.1
typing_extensions==4.6.3
urllib3==1.26.16
Werkzeug==2.3.6
yarg==0.1.9
yarl==1.9.2
zipp==3.15.0
